Apply API and unit test review conventions

+++ samples/DrillPress.SampleRules/Program.cs
@@ -1,4 +1,4 @@
 using DrillPress.Engine;
 using DrillPress.SampleRules;
 
-return await RuleApplication.RunAsync(SampleRuleSet.Create(), args);
+return (int)await RuleApplication.RunAsync(SampleRuleSet.Create(), args);
